support environment variables
My prometheus and grafana are in a docker compose setup, so eventually i'll move this into a container too.
A common pattern in docker compose is setting options via an environment variable so i want to support that
It was either that or support an ini config file.

But having it an environment means the container won't need any volume passed into it