fix production metrics removing negatives.
The production was showing less than zero.  My guess is either, it can't
help but include the energy required to run the envoy itself and 0 -
cost = < 0.

Or there is a margin or error of a few watts.

In a blog post I read some webui's hide this by setting it to zero, mine
on the other hand shows the below zero value so :shrug: